requirements.txt
google-genai
picamera2 